YoutubeGenerationProject/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

youtube-dl
azure-storage-blob
azure-functions
praw
azure-identity
azure-keyvault-secrets
certifi
chardet
decorator
ez-setup
idna
imageio
moviepy
numpy
Pillow
prawcore
pytube
requests
tqdm
update-checker
urllib3
bs4